Autenticação com Spring Security

+++ src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -1,12 +1,17 @@
 package br.com.alura.forum.controller;
 
 import java.net.URI;
-import java.util.List;
+import java.util.Optional;
 
 import javax.validation.Valid;
-import javax.websocket.server.PathParam;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.Cacheable;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort.Direction;
+import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -16,6 +21,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.util.UriComponentsBuilder;
 
@@ -40,46 +46,68 @@ public class TopicosController {
 	private CursoRepository cursoRepository;
 
 	@GetMapping
-	public List<TopicoDto> lista(String nomeCurso) {
-
+	@Cacheable(value = "listaTopicos")
+	public Page<TopicoDto> lista(@RequestParam(required = false) String nomeCurso, @PageableDefault(page=0, size=10, sort="id", direction = Direction.DESC) Pageable paginacao) {
 		if (nomeCurso == null) {
-			List<Topico> topicos = topicoRepository.findAll();
+			Page<Topico> topicos = topicoRepository.findAll(paginacao);
 			return TopicoDto.converter(topicos);
 		} else {
-			List<Topico> topicos = topicoRepository.findByCurso_Nome(nomeCurso);
+			Page<Topico> topicos = topicoRepository.findByCurso_Nome(nomeCurso, paginacao);
 			return TopicoDto.converter(topicos);
 		}
 	}
 
 	@PostMapping
 	@Transactional
+	@CacheEvict(value="listaTopicos", allEntries=true)
 	public ResponseEntity<TopicoDto> cadastrar(@RequestBody @Valid TopicoForm form, UriComponentsBuilder uriBuilder) {
 		Topico topico = form.converter(cursoRepository);
 		topicoRepository.save(topico);
-		
+
 		URI uri = uriBuilder.path("/topicos/{id}").buildAndExpand(topico.getId()).toUri();
 		return ResponseEntity.created(uri).body(new TopicoDto(topico));
 	}
-	
+
 	@GetMapping("/{id}")
-	public DetalhesTopicoDto detalhar(@PathVariable Long id){
-		Topico topico = topicoRepository.getOne(id);
-		return new DetalhesTopicoDto(topico);
+	public ResponseEntity<DetalhesTopicoDto> detalhar(@PathVariable Long id) {
+		Optional<Topico> topico = topicoRepository.findById(id);
+
+		if (topico.isPresent()) {
+			return ResponseEntity.ok(new DetalhesTopicoDto(topico.get()));
+		}
+
+		return ResponseEntity.notFound().build();
 	}
-	
+
 	@PutMapping("{id}")
 	@Transactional
-	public ResponseEntity<TopicoDto> atualizarTopico(@PathVariable Long id, @RequestBody @Valid AtualizacaoTopicoForm form, UriComponentsBuilder uriBuilder) {
-		Topico topico = form.atualizar(id, topicoRepository);
-		return ResponseEntity.ok(new TopicoDto(topico));
+	@CacheEvict(value="listaTopicos", allEntries=true)
+	public ResponseEntity<TopicoDto> atualizarTopico(@PathVariable Long id,
+			@RequestBody @Valid AtualizacaoTopicoForm form, UriComponentsBuilder uriBuilder) {
+
+		Optional<Topico> topicoOptional = topicoRepository.findById(id);
+
+		if (topicoOptional.isPresent()) {
+			Topico topico = form.atualizar(id, topicoRepository);
+			return ResponseEntity.ok(new TopicoDto(topico));
+		}
+
+		return ResponseEntity.notFound().build();
+
 	}
-	
+
 	@DeleteMapping("/{id}")
 	@Transactional
-	public ResponseEntity<?> remover(@PathVariable Long id){
-		topicoRepository.deleteById(id);
-		
-		return ResponseEntity.ok().build();
+	@CacheEvict(value="listaTopicos", allEntries=true)
+	public ResponseEntity<?> remover(@PathVariable Long id) {
+
+		Optional<Topico> topico = topicoRepository.findById(id);
+
+		if (topico.isPresent()) {
+			topicoRepository.deleteById(id);
+			return ResponseEntity.ok().build();
+		}
+		return ResponseEntity.notFound().build();
 	}
 
 	// Metodo de Teste
